requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
ansible-compat==25.8.1
    # via ansible-lint
ansible-core==2.16.14
    # via
    #   -r requirements.in
    #   ansible-compat
    #   ansible-lint
ansible-lint==25.9.1
    # via -r requirements.in
ansible-runner==2.4.1
    # via -r requirements.in
astroid==3.3.11
    # via pylint
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
azure-common==1.1.28
    # via azure-mgmt-network
azure-core==1.35.1
    # via
    #   azure-identity
    #   azure-kusto-data
    #   azure-mgmt-core
    #   azure-storage-blob
    #   azure-storage-queue
    #   msrest
azure-identity==1.25.0
    # via
    #   -r requirements.in
    #   azure-kusto-data
azure-kusto-data==5.0.5
    # via
    #   -r requirements.in
    #   azure-kusto-ingest
azure-kusto-ingest==5.0.5
    # via -r requirements.in
azure-mgmt-compute==37.0.0
    # via -r requirements.in
azure-mgmt-core==1.6.0
    # via
    #   azure-mgmt-compute
    #   azure-mgmt-network
azure-mgmt-network==29.0.0
    # via -r requirements.in
azure-storage-blob==12.23.0
    # via
    #   -r requirements.in
    #   azure-kusto-ingest
azure-storage-queue==12.12.0
    # via
    #   -r requirements.in
    #   azure-kusto-ingest
black==25.9.0
    # via
    #   -r requirements.in
    #   ansible-lint
bracex==2.6
    # via wcmatch
certifi==2025.8.3
    # via
    #   msrest
    #   requests
cffi==2.0.0
    # via
    #   ansible-lint
    #   cryptography
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   -r requirements.in
    #   black
coverage[toml]==7.10.7
    # via
    #   -r requirements.in
    #   pytest-cov
cryptography==46.0.2
    # via
    #   ansible-core
    #   ansible-lint
    #   azure-identity
    #   azure-storage-blob
    #   azure-storage-queue
    #   msal
    #   pyjwt
dill==0.4.0
    # via pylint
distro==1.9.0
    # via ansible-lint
exceptiongroup==1.3.0
    # via pytest
filelock==3.19.1
    # via ansible-lint
idna==3.10
    # via requests
ijson==3.4.0
    # via azure-kusto-data
importlib-metadata==8.7.0
    # via ansible-lint
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-mgmt-network
    #   azure-storage-blob
    #   azure-storage-queue
    #   msrest
isort==6.1.0
    # via pylint
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   ansible-core
jmespath==1.0.1
    # via -r requirements.in
jsonschema==4.25.1
    # via
    #   ansible-compat
    #   ansible-lint
jsonschema-specifications==2025.9.1
    # via jsonschema
lockfile==0.12.2
    # via python-daemon
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
msal==1.34.0
    # via
    #   azure-identity
    #   azure-kusto-data
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
msrest==0.7.1
    # via azure-mgmt-compute
mypy-extensions==1.1.0
    # via black
numpy==2.2.6
    # via
    #   -r requirements.in
    #   pandas
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   ansible-compat
    #   ansible-core
    #   ansible-lint
    #   ansible-runner
    #   black
    #   pytest
pandas==2.3.3
    # via -r requirements.in
pathspec==0.12.1
    # via
    #   ansible-lint
    #   black
    #   yamllint
pexpect==4.9.0
    # via ansible-runner
platformdirs==4.4.0
    # via
    #   black
    #   pylint
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
ptyprocess==0.7.0
    # via pexpect
pycparser==2.23
    # via cffi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
pylint==3.3.8
    # via -r requirements.in
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==7.0.0
    # via -r requirements.in
pytest-mock==3.15.1
    # via -r requirements.in
python-daemon==3.1.2
    # via ansible-runner
python-dateutil==2.9.0.post0
    # via
    #   azure-kusto-data
    #   pandas
pytokens==0.1.10
    # via black
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   ansible-compat
    #   ansible-core
    #   ansible-lint
    #   ansible-runner
    #   yamllint
referencing==0.36.2
    # via
    #   ansible-lint
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   azure-core
    #   azure-kusto-data
    #   msal
    #   msrest
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via msrest
resolvelib==1.0.1
    # via ansible-core
rich==14.1.0
    # via -r requirements.in
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.15
    # via ansible-lint
ruamel-yaml-clib==0.2.14
    # via
    #   ansible-lint
    #   ruamel-yaml
six==1.17.0
    # via
    #   azure-core
    #   python-dateutil
subprocess-tee==0.4.2
    # via
    #   ansible-compat
    #   ansible-lint
tenacity==9.1.2
    # via azure-kusto-ingest
tomli==2.2.1
    # via
    #   black
    #   coverage
    #   pylint
    #   pytest
tomlkit==0.13.3
    # via pylint
typing-extensions==4.15.0
    # via
    #   astroid
    #   azure-core
    #   azure-identity
    #   azure-mgmt-compute
    #   azure-mgmt-network
    #   azure-storage-blob
    #   azure-storage-queue
    #   black
    #   cryptography
    #   exceptiongroup
    #   referencing
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
wcmatch==10.1
    # via ansible-lint
yamllint==1.37.1
    # via ansible-lint
zipp==3.23.0
    # via importlib-metadata
